requirements_bot.txt
discord.py
azure-identity
azure-mgmt-containerinstance
aiohttp